+++ REST/src/main/java/com/study/controller/StudentController.java
@@ -1,6 +1,7 @@
 package com.study.controller;
 
 import com.study.models.Student;
+
 import com.study.repos.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -28,12 +29,36 @@ public class StudentController {
 
     @PostMapping("/student")
     public @ResponseBody Student createStudent(@RequestBody Student student) {
-        return studentRepository.save(student);
+        var student_ = studentRepository.save(student);
+        return student_;
     }
 
     @DeleteMapping("/student/{id_card}")
     public @ResponseBody void deleteStudent(@PathVariable int id_card) {
-        studentRepository.deleteById(id_card);
+        var student = studentRepository.findById(id_card);
+        if (student.isPresent()){
+            studentRepository.deleteById(id_card);
+        }
+    }
+
+    @PutMapping("/student/{id}")
+    public @ResponseBody Student updateStudent(@RequestBody Student newStudent, @PathVariable Integer id) {
+        return studentRepository.findById(id)
+                .map(student -> {
+                    if (newStudent.getAvg_score() != 0) {
+                        student.setAvg_score(newStudent.getAvg_score());
+                    }
+                    if (newStudent.getId_dorm() != 0) {
+                        student.setId_dorm(newStudent.getId_dorm());
+                    }
+                    if (newStudent.getYearEducation() != 0) {
+                        student.setYearEducation(newStudent.getYearEducation());
+                    }
+                    return studentRepository.save(student);
+                })
+                .orElseGet(() -> {
+                    return studentRepository.save(newStudent);
+                });
     }
 }
